🚚 chore: rename usuarioRoute to usuariosRoute
- Deleted `usuarioRoute.js` and updated `usuariosRoute.js` for consistent naming and improved route management.
