+++ api/routes/usuariosRoute.js
@@ -1,17 +1,17 @@
-const { Router } = require('express')
+const { Router }  = require('express')
 const UsuarioController = require('../controllers/usuarioController')
-const autenticado = require('../middleware/autenticado')
+const autenticado = require('../midlleware/autenticado')
 
-const router = Router()
-
-router.use(autenticado)
+const router = Router();
 
 router
-    .post('/usuarios', UsuarioController.cadastrar)
-    .get('/usuarios', UsuarioController.buscarTodosUsuarios)
-    .get('/usuarios/id/:id', UsuarioController.buscarUsuarioPorId)
-    .put('/usuarios/id/:id', UsuarioController.editarUsuario)
-    .delete('/usuarios/id/:id', UsuarioController.deletarUsuario)
-
+  .post('/usuarios', UsuarioController.cadastrar)
+router
+  .use(autenticado)
+router
+  .get('/usuarios', UsuarioController.buscarTodosUsuarios)
+  .get('/usuarios/:id', UsuarioController.buscarUsuarioPorId)
+  .put('/usuarios/:id', UsuarioController.editarUsuario)
+  .delete('/usuarios/:id', UsuarioController.deletarUsuario)
 
 module.exports = router
